+++ src/__init__.py
@@ -1,5 +1,12 @@
 """Package: drug-interaction-db"""
 
+from src.alternative_suggester import (
+    THERAPEUTIC_CLASSES,
+    AlternativeSuggestion,
+    list_classes,
+    suggest_alternatives,
+    suggestions_to_dataframe,
+)
 from src.cyp450_tagger import (
     KNOWN_ENZYMES,
     EnzymeTag,
@@ -17,15 +24,20 @@ from src.polypharmacy_risk_scorer import (
 )
 
 __all__ = [
+    "AlternativeSuggestion",
     "DrugInteractionDB",
     "EnzymeTag",
     "KNOWN_ENZYMES",
     "PairScore",
     "RiskReport",
+    "THERAPEUTIC_CLASSES",
     "extract_enzymes",
     "filter_by_enzyme",
+    "list_classes",
     "report_to_dataframe",
     "score_regimen",
+    "suggest_alternatives",
+    "suggestions_to_dataframe",
     "summarise_by_enzyme",
     "tag_interactions",
 ]